+++ src/main/nextflow/co2footprint/CO2FootprintFactory.groovy
@@ -1,26 +1,17 @@
 package nextflow.co2footprint
 
+import groovy.transform.CompileStatic
+import groovy.transform.PackageScope
+import groovy.transform.PackageScopeTarget
+import groovy.util.logging.Slf4j
+import nextflow.Session
 import nextflow.co2footprint.DataContainers.CIDataMatrix
 import nextflow.co2footprint.DataContainers.TDPDataMatrix
 import nextflow.co2footprint.Logging.LoggingAdapter
-
-import java.nio.file.Path
-import java.nio.file.Paths
-
-import groovy.transform.PackageScope
-import groovy.transform.PackageScopeTarget
-import groovy.transform.CompileStatic
-
-import groovy.util.logging.Slf4j
-import org.slf4j.LoggerFactory
-import ch.qos.logback.classic.LoggerContext
-
-import nextflow.Session
 import nextflow.trace.TraceObserver
 import nextflow.trace.TraceObserverFactory
 
-import java.util.jar.Manifest
-
+import java.nio.file.Paths
 
 /**
  * Factory class for creating the CO2Footprint trace observer.