+++ src/main/nextflow/co2footprint/Config/FileSubConfig.groovy
@@ -17,6 +17,7 @@ import java.lang.Boolean
 class FileSubConfig extends BaseConfig {
     private String name
     private String timestamp = TraceHelper.launchTimestampFmt()
+    private String ending = 'txt'
 
     private void defineParameters() {
         // Name, description, default value or function, return type, additional allowed types
@@ -25,18 +26,20 @@ class FileSubConfig extends BaseConfig {
         )
         defineParameter(
                 'file', 'Path to the output file',
-                "co2footprint_${name}_${timestamp}.txt", String, Set.of(GString, Path)
+                "co2footprint_${name}_${timestamp}.${ending}", String, Set.of(GString, Path)
         )
     }
 
     /**
      * Load a nested map configuration and set up defaults and fallbacks.
      *
-     * @param subConfigName  Name for this sub-config - used to get a default file name for this output file
      * @param configMap      Map of configuration values (from Nextflow config)
+     * @param subConfigName  Name for this sub-config - used to get a default file name for this output file
+     * @param fileEnding     Ending type of the file
      */
-    FileSubConfig(String subConfigName, Map<String, Object> configMap) {
+    FileSubConfig(Map<String, Object> configMap, String subConfigName, String fileEnding='txt') {
         this.name = subConfigName
+        this.ending = fileEnding
 
         // Define the possible parameters of the configuration
         defineParameters()